[#13] Lunch 데이터 그리드 보기 좋게 하기
* 기존 DateLoader에 선언되어있던 getDateFromDateOnlyStr()을 FormatUtil.ts 파일로 이동
* prettier 설정 추가 ("arrowParens": "avoid")

** cellStyle 대신 type: 'rightAligned', type: 'numericColumn'도 사용할 수 있는데 헤더까지 정렬한다.
