+++ lt-webapp/src/App.vue
@@ -22,12 +22,12 @@
         font-family: Avenir, Helvetica, Arial, sans-serif;
         -webkit-font-smoothing: antialiased;
         -moz-osx-font-smoothing: grayscale;
-        text-align: center;
         color: #2c3e50;
     }
 
     #nav {
         padding: 30px;
+        text-align: center;
 
         a {
             font-weight: bold;
