+++ lt-webapp/src/common/DataLoader.ts
@@ -1,4 +1,5 @@
 import datasource from "@/data/datasource.json";
+import * as FormatUtil from "@/common/FormatUtil";
 import { Lunch, Place } from "@/common/TypeDef";
 
 const NullString = "";
@@ -9,26 +10,6 @@ function getValueWithDefault(value: any, defaultValue: any): any {
     if (value === null || value === undefined || value === NullString)
         return defaultValue;
     return value;
-}
-
-function substringAsNumber(str: string, start: number, end: number): number {
-    return Number(str.substring(start, end));
-}
-
-// TODO Type을 define하는 방법도 있는지 찾아보자.
-function getDateFromDateOnlyStr(dateOnlyStr: string): Date {
-    // prettier-ignore
-    if (dateOnlyStr.length !== 8)
-        return new Date(-1);
-
-    return new Date(
-        substringAsNumber(dateOnlyStr, 0, 4),
-        substringAsNumber(dateOnlyStr, 4, 6),
-        substringAsNumber(dateOnlyStr, 6, 8),
-        0,
-        0,
-        0
-    );
 }
 // 여기까지 위치가 애매하다. 일단은 여기서 쓰니까 여기 둔다.
 
@@ -42,7 +23,7 @@ export const lunch: Array<Lunch> = datasource.lunch.map(l => ({
     id: l.no,
     menu: l.name,
     price: getValueWithDefault(l.price, 0),
-    date: new Date(getDateFromDateOnlyStr(l.date)),
+    date: new Date(FormatUtil.getDateFromDateOnlyStr(l.date)),
     placeId: l.placeNo
 }));
 Object.freeze(lunch);
